* commit '2a6b25f31814f372201516521a2479ee4de2d09e':
  Overhaul wvstreams build system from Makefiles to redo.
  Remove support for boost/function.hpp in favour of tr1/functional.
  wvstreams.supp: remove as libc6 and openssl are nowadays valgrind-clean.
  Fix compilation under mingw64; wvtestmain runs under wine64.

+++ utils/t/wvcallback.t.cc
@@ -20,7 +20,7 @@ typedef wv::function<A(void*)> A3Callback;
 
 static A bunk(const A &a, void *userdata)
 {
-    long incr = (long)userdata;
+    long incr = (long)(long long)userdata;
     return A(a.x+incr, a.y+incr*2);
 }
